// src/services/entities/app/legalPerson/school/enums.ts
export enum EducationLevelEnum {
  EducacaoInfantil = 'EducacaoInfantil',
  EnsinoFundamental = 'EnsinoFundamental',
  EnsinoMedio = 'EnsinoMedio',
  Graduacao = 'Graduacao',
  GraduacaoTecnologica = 'GraduacaoTecnologica',
  PosGraduacao = 'PosGraduacao',
  Mestrado = 'Mestrado',
  Doutorado = 'Doutorado',
  PosDoutorado = 'PosDoutorado',
  Extensao = 'Extensao',
  Profissionalizante = 'Profissionalizante',
  EducacaoEmpresarial = 'EducacaoEmpresarial',
}